feat: mobile calendar view 대응
