+++ client/src/domains/eventPage/components/CalendarView/mobile/MobileCalendarView.module.css
@@ -1,0 +1,18 @@
+.wrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  margin-top: 50px;
+}
+
+.hand {
+  font-size: 50px;
+  margin-bottom: 16px;
+}
+
+.title {
+  font-size: 16px;
+  margin-bottom: 4px;
+  font-weight: bold;
+}
