+++ client/src/domains/eventPage/components/MainEvents/styled.js
@@ -17,6 +17,7 @@ export const Container = styled.div`
   @media screen and (max-width: 768px) {
     padding: 0;
     margin-top: 25px;
+    min-height: 0;
   }
 `;
